Update imports to reflect new location of StatusNotification - Moved from src/components to src/features/notifications
